checks/promql/series: make lookbackrange user configurable
If the instance stores less than 7 days of data, which is common for
thanos, there is no reason to try to query 7 days into the past.

This setting can also help reduce the query load when enabling pint on
systems with a lot of broken rules. As each test condition queries the
whole range a single broken rule can lead to 100s of queries. Multiplied
by 100s of broken rules and a regular `watch` interval this adds up.